refactor: move game coponent to its own directory
