+++ web-app/src/components/home.tsx
@@ -1,5 +1,5 @@
 import { generatePuzzle } from "../shared/generate_puzzle";
-import { Game } from "./game";
+import { Game } from "./game/game";
 
 function Home() {
   return (
